more clean up

+++ Frost/src/Event/Event.c
@@ -8,12 +8,12 @@ void EventReset(Event* e)
 	e->handled = 0;
 }
 
-int EventCheckType(const Event* e, EventType type)
+uint8_t EventCheckType(const Event* e, EventType type)
 {
 	return !e->handled && e->type == type;
 }
 
-int EventMouseButton(const Event* e)
+int32_t EventMouseButton(const Event* e)
 {
 	if (e->handled) return MOUSE_BUTTON_UNKNOWN;
 
@@ -23,21 +23,21 @@ int EventMouseButton(const Event* e)
 	return MOUSE_BUTTON_UNKNOWN;
 }
 
-int EventMouseButtonPressed(const Event* e)
+int32_t EventMouseButtonPressed(const Event* e)
 {
 	if (EventCheckType(e, EVENT_MOUSE_BUTTON_PRESSED)) return e->mousebutton.buttoncode;
 
 	return MOUSE_BUTTON_UNKNOWN;
 }
 
-int EventMouseButtonReleased(const Event* e)
+int32_t EventMouseButtonReleased(const Event* e)
 {
 	if (EventCheckType(e, EVENT_MOUSE_BUTTON_RELEASED)) return e->mousebutton.buttoncode;
 
 	return MOUSE_BUTTON_UNKNOWN;
 }
 
-int EventKey(const Event* e)
+int32_t EventKey(const Event* e)
 {
 	if (e->handled) return KEY_UNKNOWN;
 
@@ -47,14 +47,14 @@ int EventKey(const Event* e)
 	return KEY_UNKNOWN;
 }
 
-int EventKeyPressed(const Event* e)
+int32_t EventKeyPressed(const Event* e)
 {
 	if (EventCheckType(e, EVENT_KEY_PRESSED)) return e->key.keycode;
 
 	return KEY_UNKNOWN;
 }
 
-int EventKeyReleased(const Event* e)
+int32_t EventKeyReleased(const Event* e)
 {
 	if (EventCheckType(e, EVENT_KEY_RELEASED)) return e->key.keycode;
