+++ Frost/src/Event/EventHandler.h
@@ -3,12 +3,12 @@
 
 #include "Event.h"
 
-int EventHandlerInit(size_t queue_size, void (*callback)(void*, Event));
+uint8_t EventHandlerInit(size_t queue_size, void (*callback)(void*, Event));
 void EventHandlerDestroy();
 
-void EventHandlerThrowWindowEvent(EventType type, int width, int height);
-void EventHandlerThrowKeyEvent(EventType type, int keycode, int repeatcount);
-void EventHandlerThrowMouseButtonEvent(EventType type, int button);
+void EventHandlerThrowWindowEvent(EventType type, uint32_t width, uint32_t height);
+void EventHandlerThrowKeyEvent(EventType type, int32_t keycode, uint32_t repeatcount);
+void EventHandlerThrowMouseButtonEvent(EventType type, int32_t button);
 void EventHandlerThrowMouseEvent(EventType type, float x, float y);
 void EventHandlerThrowConsoleEvent(EventType type, const char* cmd);
